Project/CryptoTradingBot/requirements.txt
# Crypto Trading Bot - Python Dependencies

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Advanced ML algorithms
xgboost>=2.0.0
lightgbm>=4.0.0
optuna>=3.0.0  # Hyperparameter optimization
shap>=0.42.0  # Feature importance analysis

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Database
# sqlite3  # Built-in with Python

# API and web framework
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0

# Cryptocurrency exchange connectivity
ccxt>=4.0.0

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# Logging and monitoring
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Environment management
python-dotenv>=1.0.0

# Date/time handling
python-dateutil>=2.8.0

# HTTP client
httpx>=0.24.0

# Data analysis
ydata-profiling>=4.0.0  # Successor to pandas-profiling (Python 3.11 compatible)

# Performance monitoring
psutil>=5.9.0

# Security
cryptography>=41.0.0

# Async support
asyncio-mqtt>=0.13.0

# Workflow orchestration
prefect>=2.14.0

# Additional utilities
rich>=13.0.0
click>=8.1.0
tqdm>=4.65.0
